+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ Release Notes
 * ExpressRoute: create authorization keys on newly created circuits.
 * ServiceBus: Allow Service Bus Queues/Topics/Subscriptions to be linked to unmanaged namespaces
 * ServiceBus: Allow adding custom dependencies to Subscriptions
+* WebApp/Functions: Adds 'ftp_state' for controlling FTP access for deployments.
 
 ## 1.6.18
 * Resource Groups: Add support for multiple nested deployments targetting the same resource group